requirements.txt
Django==3.1.1
django-mailgun==0.9.1
gunicorn==20.0.4
Markdown==3.2.2
whitenoise==5.1.0
Pillow==7.2.0
djangorestframework==3.11.0
dj-database-url==0.5.0
psycopg2==2.8.5
django-dotenv==1.4.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.5.0
django-googledrive-storage==1.5.2
